Trigger the firstrunwizard on any app
